Changes to code stucture - still needs work
