+++ Admin-app/src/components/ChatList/ChatList.js
@@ -7,10 +7,24 @@ import ListItemAvatar from '@mui/material/ListItemAvatar';
 import Avatar from '@mui/material/Avatar';
 import Typography from '@mui/material/Typography';
 import { useAuthState } from 'react-firebase-hooks/auth';
-import { doc, collection, query,onSnapshot, orderBy, limit} from "firebase/firestore";
+import { doc, collection, query,onSnapshot, orderBy, limit, setDoc} from "firebase/firestore";
 import { db, auth } from "../../backend/firebase";
 import { useEffect, useState } from "react";
 import FlagIcon from "@mui/icons-material/Flag";
+import { red } from '@mui/material/colors';
+import {
+    togglePriorityFlag
+} from "../../backend/firebasePatientUtilities";
+import './ChatList.css'
+
+// priority flag with DB
+// function onFlagClick(id) {
+//     togglePriorityFlag(id).then(
+//       (newPatientInfo) =>
+//         newPatientInfo &&
+//         setPriorityFlag(newPatientInfo.flaggedPriority === "1")
+//     );
+//   }
 
 function stringToColor(string) {
     let hash = 0;
@@ -50,6 +64,7 @@ export default function ChatList(props) {
     const adminRef = doc(db, "Admin", user.email)
     const messageRef = collection(adminRef, "Clients")
     const q = query(messageRef)
+    const [priorityFlag, setPriorityFlag] = useState(false);
 
     // This hook shows all the clients that belong to the doctor as a list. 
     useEffect(() => {
@@ -79,18 +94,44 @@ function PatientsList(props) {
     
     const clientRef = doc(db, "Client", name)
     const messageRef = collection(clientRef, "Messages")
+    // const countCol = collection(clientRef, "Counter")
+    // const countRef = doc(countCol, "counter")
     const q = query(messageRef, orderBy('timestamp', 'desc'), limit(1));
     const [lastMessage, setLastMessage] = useState('');
+    const [flag, setFlag] = useState(false);
+    // const [messageCount, setMessageCount] = useState(0);
 
     // This hook is used to show the lastest message sent on the database. 
     useEffect(() => {
+
         onSnapshot(q, (doc) => {
             setLastMessage(doc.docs.map(doc=> ({
-                message: doc.data().message,
+                message: doc.data()?.message,
             })))
         })
+
+        onSnapshot(clientRef, (doc) => {
+            setFlag(doc.data()?.flag)
+        })
+
+        // onSnapshot(countRef, (doc) => {
+        //     setMessageCount(doc.data()?.counterDoc)
+        // })
     // eslint-disable-next-line
     }, [])
+
+    const handleClick = async (event) => {
+        // event.preventDefault();
+        // if(flag){
+        //     await updateDoc(clientRef, {
+        //         flag: false,
+        //     })
+        // }else{
+        //     await updateDoc(clientRef, {
+        //         flag: true,
+        //     })
+        // }
+    }
 
     if(!lastMessage[0]){
         return(
@@ -103,7 +144,14 @@ function PatientsList(props) {
                     primary= {
                         <React.Fragment>
                             <Typography style={{ maxWidth: '100px',}} color="var(--text-primary)"> {name} </Typography>
-                            {<FlagIcon label="primary" color="primary" variant="outlined" />}
+                            {<FlagIcon label="primary" color="primary" variant="outlined"
+                            // onClick={() => {
+                            //     onFlagClick(id);
+                            //   }}
+                            //   className={
+                            //     priorityFlag ? "priority-flag clicked" : "priority-flag"
+                            //   } 
+                            />}
                         </React.Fragment>}
                     secondary={
                         <React.Fragment>
@@ -126,13 +174,14 @@ function PatientsList(props) {
                         <React.Fragment>
                             <Typography color="var(--text-primary)"> 
                                 {name} 
-                                <FlagIcon style={{ marginLeft: '10px',}} label="error" color="error" variant="outlined" />
+                                <FlagIcon style={{ marginLeft: '10px',}} label="error" color="error" variant="outlined" 
+                                    onClick={() => {handleClick()}} className={flag ? "priority-flag clicked" : "priority-flag"} />
+                                {/* <Avatar sx={{ width: 24, height: 24, bgcolor: red[500]}} >{messageCount}</Avatar> */}
                             </Typography>
                         </React.Fragment>}
                     secondary={
                         <React.Fragment>
                             <Typography style={{ maxWidth: '100px',}} color="var(--text-inactive)"> {lastMessage && lastMessage[0].message} </Typography>
-                            {/* {"last comment"} */}
                         </React.Fragment>
                     }
                 />
