+++ Client-app/src/components/Chat/Chat.css
@@ -1,7 +1,7 @@
 #messagesReceived {
   padding: 10px;
   max-width: 100vw; 
-  max-height: 72.5vh; 
+  height: 75vh; 
   overflow-y: scroll;
   display: flex;
   flex-direction: column;
@@ -35,11 +35,11 @@ li {
 
 .sent p {
   color: white;
-  background: #0b93f6;
+  background: #8f96e2;
   align-self: flex-end;
 }
 
 .received p {
-  background: #e5e5ea;
-  color: black;
+  background: #474646;
+  color: white;
 }
